[api][webui] Move Group and User query logic to it's own classes
